Updated Entities and relationships, modified BillingService

+++ Models/ConsumerTariffs/TariffInfoDto.cs
@@ -3,8 +3,8 @@
     public class TariffInfoDto
     {
         public string TariffName { get; set; }
-        public decimal BaseRate { get; set; }
-        public decimal TaxRate { get; set; }
+        public double BaseRate { get; set; }
+        public double TaxRate { get; set; }
         public List<TodRuleDto> TodRules { get; set; }
         public List<TariffSlabDto> TariffSlabs { get; set; }
     }
