+++ Models/ConsumerTariffs/TodRuleDto.cs
@@ -5,6 +5,6 @@
         public string Name { get; set; }
         public TimeOnly StartTime { get; set; }
         public TimeOnly EndTime { get; set; }
-        public decimal RatePerKwh { get; set; }
+        public double RatePerKwh { get; set; }
     }
 }
